+++ src/pages/shops/[id]/edit.tsx
@@ -1,29 +1,8 @@
-import { useRouter } from 'next/router';
 import Sidebar from '@/Components/html/Sidebar';
-import { IShop } from '@/Models/Shop';
-import { useEffect, useState } from 'react';
+import ShopId from '@/Components/single/shopId';
 
 export default function ShopEditById() {
-  const router = useRouter();
-  const id = router.query.id;
-  const [shop, setShop] = useState<IShop>();
-  useEffect(() => {
-    fetch(`/api/shop/${id}`)
-      .then((res) => res.json())
-      .then((data) => setShop(data));
-  }, []);
-
-  return (
-    <div>
-      <h1>{shop?.companyName}</h1>
-      <p>{shop?.address}</p>
-      <p>{shop?.zipCode}</p>
-      <p>{shop?.phone}</p>
-      <p>{shop?.email}</p>
-      <p>{shop?.siren}</p>
-      <p>{shop?.siret}</p>
-    </div>
-  );
+  return <ShopId />;
 }
 ShopEditById.getLayout = (page) => (
   <div className="flex">
